requirements.txt
dash==1.19.0
dash_core_components==1.15.0
dash_html_components==1.1.2
numpy==1.19.5
pandas==1.3.3
plotly==4.14.3
python-dotenv~=0.19.0
setuptools~=52.0.0
binance~=0.3
pymitter~=0.3.1
python-binance~=1.0.15
requests~=2.27.1
fastapi~=0.72.0
starlette~=0.17.1
AutoTrader~=0.0.1
uvicorn~=0.17.0